chore: Fix full auto release process
