+++ .github/workflows/release-final.yml
@@ -84,7 +84,7 @@ jobs:
           RELEASE_VERSION: ${{ inputs.version }}
           RELEASE_PREVIOUS_VERSION: ${{ inputs.previous_version }}
           DRY_RUN: ${{ inputs.dry_run }}
-          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          GH_TOKEN: ${{ secrets.RELEASE_PAT }}
           GITHUB_RUN_ID: ${{ github.run_id }}
           OPENCODE_API_KEY: ${{ secrets.OPENCODE_API_KEY }}
         run: npm run release-final
