+++ package.json
@@ -66,7 +66,8 @@
         "start-dev-container": "cd docker && docker-compose -f docker-compose-dev.yml up --force-recreate",
         "rebase-pr-to-1.23.X": "node extra/rebase-pr.js 1.23.X",
         "reset-migrate-aggregate-table-state": "node extra/reset-migrate-aggregate-table-state.js",
-        "generate-changelog": "node ./extra/generate-changelog.mjs"
+        "generate-changelog": "node ./extra/generate-changelog.mjs",
+        "create-release": "node ./extra/release/create-release.mjs"
     },
     "dependencies": {
         "@grpc/grpc-js": "~1.8.22",
